+++ controllers/productController.js
@@ -7,7 +7,7 @@ const response = require("../network/response");
 exports.listProduct = async (req, res, next) => {
   try {
     const products = await db.product.findAll();
-    response.success(req, res, products, 201);
+    response.success(req, res, products, 200);
   } catch (err) {
     console.error(chalk.red("ctr-prod-list"), err);
     response.error(req, res, err.message, 500);
@@ -22,7 +22,7 @@ exports.getProduct = async (req, res, next) => {
       },
     });
     if (product !== null) {
-      response.success(req, res, product, 201);
+      response.success(req, res, product, 200);
     } else {
       response.success(req, res, "Not found", 404);
     }
@@ -52,22 +52,33 @@ exports.insertProduc = async (req, res, next) => {
 
 exports.updatedProduct = async (req, res, next) => {
   try {
-    const { name, price, img } = req.body;
-
-    const updatedProduct = await db.product.update(
-      {
-        name,
-        price,
-        img,
+    const product = await db.product.findOne({
+      where: {
+        id: req.params.id,
       },
-      {
+    });
+    if (product) {
+      const { name, price, img } = req.body;
+      await db.product.update(
+        {
+          name,
+          price,
+          img,
+        },
+        {
+          where: {
+            id: req.params.id,
+          },
+        }
+      );
+      const productUpdated = await db.product.findOne({
         where: {
           id: req.params.id,
         },
-      }
-    );
-    if (updatedProduct.length > 0) {
-      response.success(req, res, updatedProduct, 201);
+      });
+      response.success(req, res, productUpdated, 201);
+    } else {
+      response.error(req, res, "Not found", 404);
     }
   } catch (err) {
     console.error(chalk.red("ctr-prod-updated"), err);
@@ -77,12 +88,21 @@ exports.updatedProduct = async (req, res, next) => {
 
 exports.deletedProduct = async (req, res, next) => {
   try {
-    const deletedProduct = await db.product.destroy({
+    const product = await db.product.findOne({
       where: {
         id: req.params.id,
       },
     });
-    response.success(req, res, deletedProduct, 201);
+    if (product) {
+      await db.product.destroy({
+        where: {
+          id: req.params.id,
+        },
+      });
+      response.success(req, res, "The product has been removed", 200);
+    } else {
+      response.error(req, res, "Not found", 404);
+    }
   } catch (err) {
     console.error(chalk.red("ctr-prod-delete"), err);
     response.error(req, res, err.message, 500);
